requirements.txt
requests
pyopenssl
